pyproject.toml
[project]
name = "repo-explainer"
version = "0.1.0"
description = "将 Git 仓库历史转换为通俗易懂的文档"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
